requirements.txt
azure-ai-formrecognizer
azure-core
openpyxl (excel)
numbers-parser (numbers)
PyPDF2
pillow
PyMuPDF
